Switched to AWT geom implementation for shape representation.

+++ core/src/com/solanum/hybridity/GameScreen.java
@@ -108,7 +108,7 @@ public class GameScreen implements Screen {
             }
         }
 
-        gameStage.addActor(new Seeder(ml.oX + newX, ml.oY + newY, ml.oX, ml.oY, ml));
+        gameStage.addActor(new Seeder((int) (ml.oX + newX), (int) (ml.oY + newY), ml.oX, ml.oY, ml));
 
 
     }
